refactor(frontend): Remove duplicate ProjectConfig class
