--- frontend/src/app/model/project-config.ts
@@ -1,7 +1,0 @@
-export class ProjectConfig {
-  packageName: string;
-  projectName: string;
-  projectVersion: string;
-  scenarioFileName: string;
-  decoratorClassName?: string;
-}
